requirements.txt
pynacl
base58
rich